IMPORTANT! fixed a very serious bug in write_session/cookies! detail:
previously, we're not correctly remove invalid access token in our
cookie or session when we found the access token is invalid!
thus rendering users with :auto_authorize on and :write_session on
end up with redirect loop if the user has an invalid access token.

this bug shouldn't be introduced at all, sorry folks!