adopt changes from rest-firebase (should really remove firebase!)
but i am not sure how to proceed. just remove it and update the docs?

also, this change was intended for webmock, which would look at
Content-Type and if it's json, then it would try to parse it so
that we could match with hashes instead of json (which since would be
in the string form, key order would be dependent, a bad thing for test)