pass ASYNC => true because we don't want to block the sequent requests
without ASYNC => true, only while it blocks (yields) would it make
the sequent requests, but we do want it to immediately make those
requests. this raised a question, even with ASYNC => true, we might
also want to raise an exception upon futures yielding, but for now
we didn't.

we probably need to store the exception under ASYNC, and still raise
an exception each time a future is yielding, instead of giving
unspecified behaviour that the response might have nothing.